GUACAMOLE-168: Update configure.ac
The build steps for building the x-server driver utilizes a feature from
pkg-config that was only added in version 0.28 (the `PKG_CHECK_VAR` macro).

This version of pkg-config is not available from default channels for
several linux distributions, making it difficult to build.

This removes the use of `PKG_CHECK_VAR` and instead uses the same
approach for configuring the FreeRDP module directory, allowing this to
build with older versions of pkg-config again.