GUACAMOLE-168: Include SSL support for the x-server driver
This updates the x-server driver to obey the `ENABLE_SSL` compilation
flag and sets up the driver daemon to use an SSL socket in the same
fashion as the standard daemon.

To configure the SSL certificate and key, new options from the xorg.conf
file can be specified, `SSLCertFile` and `SSLKeyFile` which are absolute
paths to the certificate and key files for the SSL context,
respectively.